api/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
